requirements.txt
python-dotenv
ffmpeg-python
redis
boto3